Fix app OOM on EC2: use float16, fix trainer tokenizer arg
Base model in float32 (3GB) + r=128 adapter (564MB) exceeded 8GB
instance RAM. Switch to float16 (~1.5GB) to fit comfortably.
Also fix Trainer tokenizer→processing_class rename and optimizer
fallback for non-CUDA.

+++ melancholic_poet_qlora/app.py
@@ -69,7 +69,7 @@ def load_base_model():
         tokenizer.pad_token = tokenizer.eos_token
     model = AutoModelForCausalLM.from_pretrained(
         BASE_MODEL,
-        torch_dtype=torch.float32,
+        torch_dtype=torch.float16,
         low_cpu_mem_usage=True,
     )
     model.eval()
@@ -85,7 +85,7 @@ def load_full_model():
         tokenizer.pad_token = tokenizer.eos_token
     model = AutoModelForCausalLM.from_pretrained(
         str(FULL_MODEL_DIR),
-        torch_dtype=torch.float32,
+        torch_dtype=torch.float16,
         low_cpu_mem_usage=True,
     )
     model.eval()
